Move functions that use DQ to a different compilation unit
