+++ src/VecSim/spaces/IP/IP_AVX512.h
@@ -14,8 +14,3 @@ double FP64_InnerProductSIMD8ExtResiduals_AVX512(const void *pVect1v, const void
 double FP64_InnerProductSIMD2Ext_AVX512_noDQ(const void *pVect1, const void *pVect2, size_t qty);
 double FP64_InnerProductSIMD2ExtResiduals_AVX512_noDQ(const void *pVect1v, const void *pVect2v,
                                                       size_t qty);
-#ifdef __AVX512DQ__
-double FP64_InnerProductSIMD2Ext_AVX512(const void *pVect1v, const void *pVect2v, size_t qty);
-double FP64_InnerProductSIMD2ExtResiduals_AVX512(const void *pVect1v, const void *pVect2v,
-                                                 size_t qty);
-#endif
