+++ src/VecSim/spaces/L2/L2_AVX512.h
@@ -12,8 +12,3 @@ double FP64_L2SqrSIMD8ExtResiduals_AVX512(const void *pVect1v, const void *pVect
 double FP64_L2SqrSIMD2Ext_AVX512_noDQ(const void *pVect1v, const void *pVect2v, size_t qty);
 double FP64_L2SqrSIMD2ExtResiduals_AVX512_noDQ(const void *pVect1v, const void *pVect2v,
                                                size_t qty);
-
-#ifdef __AVX512DQ__
-double FP64_L2SqrSIMD2Ext_AVX512(const void *pVect1v, const void *pVect2v, size_t qty);
-double FP64_L2SqrSIMD2ExtResiduals_AVX512(const void *pVect1v, const void *pVect2v, size_t qty);
-#endif
