+++ src/VecSim/spaces/L2/L2_AVX512DQ.h
@@ -1,0 +1,5 @@
+#pragma once
+#include <cstdlib>
+
+double FP64_L2SqrSIMD2Ext_AVX512(const void *pVect1v, const void *pVect2v, size_t qty);
+double FP64_L2SqrSIMD2ExtResiduals_AVX512(const void *pVect1v, const void *pVect2v, size_t qty);
